requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
matplotlib>=3.8.2
pillow>=10.0.0
numpy>=1.24.0
